+++ mkFit/CandCloner.cc
@@ -89,7 +89,7 @@ void CandCloner::ProcessSeedRange(int is_beg, int is_end)
 
       //cands[ m_start_seed + is ].swap(cv); // segfaulting w/ backwards fit input tracks -- using loop below now
       cands[m_start_seed+is].resize(cv.size());
-      for (int ii = 0; ii < cv.size(); ++ii)
+      for (size_t ii = 0; ii < cv.size(); ++ii)
       {
 	memcpy( & cands[m_start_seed+is][ii], & cv[ii], sizeof(Track));
       }
